requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
datasets==2.12.0
dill==0.3.6
fastapi==0.95.2
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.5.0
h11==0.14.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0
numpy==1.24.1
packaging==23.1
pandas==2.0.1
Pillow==9.3.0
protobuf==3.20.0
pyarrow==12.0.0
pydantic==1.10.8
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.28.1
responses==0.18.0
safetensors==0.3.1
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers @ git+https://github.com/huggingface/transformers.git@f67dac97bdc63874f2288546b3fa87e69d2ea1c8
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.22.0
xxhash==3.2.0
yarl==1.9.2
